Rename the collector package to detail
