+++ src/main/java/com/blackducksoftware/integration/hub/notification/content/detail/PolicyOverrideDetailFactory.java
@@ -21,31 +21,25 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.blackducksoftware.integration.hub.notification.content.collector;
+package com.blackducksoftware.integration.hub.notification.content.detail;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 
 import com.blackducksoftware.integration.hub.notification.content.NotificationContent;
 import com.blackducksoftware.integration.hub.notification.content.NotificationContentDetail;
-import com.blackducksoftware.integration.hub.notification.content.RuleViolationClearedNotificationContent;
+import com.blackducksoftware.integration.hub.notification.content.PolicyOverrideNotificationContent;
 
-public class RuleViolationClearedDetailFactory extends NotificationDetailFactory {
+public class PolicyOverrideDetailFactory extends NotificationDetailFactory {
 
     @Override
     public List<NotificationContentDetail> createDetails(final NotificationContent notificationContent) {
-        final RuleViolationClearedNotificationContent content = (RuleViolationClearedNotificationContent) notificationContent;
-        final Map<String, String> uriToName = content.policyInfos.stream().collect(Collectors.toMap(policyInfo -> policyInfo.policy, policyInfo -> policyInfo.policyName));
+        final PolicyOverrideNotificationContent content = (PolicyOverrideNotificationContent) notificationContent;
         final List<NotificationContentDetail> details = new ArrayList<>();
-        content.componentVersionStatuses.forEach(componentVersionStatus -> {
-            componentVersionStatus.policies.forEach(policyUri -> {
-                final String policyName = uriToName.get(policyUri);
-                details.add(NotificationContentDetail.createDetail(content, content.projectName, content.projectVersionName, content.projectVersion, componentVersionStatus.componentName,
-                        componentVersionStatus.component, componentVersionStatus.componentVersionName, componentVersionStatus.componentVersion, policyName,
-                        policyUri, null, componentVersionStatus.componentIssueLink, null));
-            });
+        content.policyInfos.forEach(policyInfo -> {
+            details.add(NotificationContentDetail.createDetail(content, content.projectName, content.projectVersionName, content.projectVersion, content.componentName,
+                    content.component, content.componentVersionName, content.componentVersion, policyInfo.policyName,
+                    policyInfo.policy, null, null, null));
         });
         return details;
     }
