+++ src/main/java/com/blackducksoftware/integration/hub/notification/content/detail/VulnerabilityDetailFactory.java
@@ -21,19 +21,26 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.blackducksoftware.integration.hub.notification.content.collector;
+package com.blackducksoftware.integration.hub.notification.content.detail;
 
-import java.util.Collections;
+import java.util.ArrayList;
 import java.util.List;
 
 import com.blackducksoftware.integration.hub.notification.content.NotificationContent;
 import com.blackducksoftware.integration.hub.notification.content.NotificationContentDetail;
+import com.blackducksoftware.integration.hub.notification.content.VulnerabilityNotificationContent;
 
-public class LicenseLimitDetailFactory extends NotificationDetailFactory {
+public class VulnerabilityDetailFactory extends NotificationDetailFactory {
 
     @Override
     public List<NotificationContentDetail> createDetails(final NotificationContent notificationContent) {
-        return Collections.emptyList();
+        final VulnerabilityNotificationContent content = (VulnerabilityNotificationContent) notificationContent;
+        final List<NotificationContentDetail> details = new ArrayList<>();
+        content.affectedProjectVersions.forEach(projectVersion -> {
+            details.add(NotificationContentDetail.createDetail(content, projectVersion.projectName, projectVersion.projectVersionName, projectVersion.projectVersion, content.componentName,
+                    null, content.versionName, content.componentVersion, null, null,
+                    content.componentVersionOriginName, projectVersion.componentIssueUrl, content.componentVersionOriginId));
+        });
+        return details;
     }
-
 }
